changed plotter to a initializable class
